# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusOS shared API for complaint routing and campus operations
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ComplaintStatus = typeof ComplaintStatus[keyof typeof ComplaintStatus];


export const ComplaintStatus = {
  open: 'open',
  assigned: 'assigned',
  resolved: 'resolved',
} as const;

export type ComplaintCategory = typeof ComplaintCategory[keyof typeof ComplaintCategory];


export const ComplaintCategory = {
  facilities: 'facilities',
  academics: 'academics',
  safety: 'safety',
  it: 'it',
  dining: 'dining',
} as const;

export type ComplaintSeverity = typeof ComplaintSeverity[keyof typeof ComplaintSeverity];


export const ComplaintSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Complaint {
  id: string;
  title: string;
  description: string;
  category: ComplaintCategory;
  severity: ComplaintSeverity;
  status: ComplaintStatus;
  location: string;
  createdAt: string;
  priority: number;
  priorityLabel: string;
}

export interface CreateComplaintInput {
  /**
     * @minLength 5
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description: string;
  category: ComplaintCategory;
  severity: ComplaintSeverity;
  /**
     * @minLength 2
     * @maxLength 160
     */
  location: string;
}

export type ActivityTone = typeof ActivityTone[keyof typeof ActivityTone];


export const ActivityTone = {
  indigo: 'indigo',
  cyan: 'cyan',
  amber: 'amber',
  rose: 'rose',
  emerald: 'emerald',
} as const;

export interface Activity {
  id: string;
  title: string;
  detail: string;
  time: string;
  tone: ActivityTone;
}

export interface BreakdownItem {
  label: string;
  value: number;
  color: string;
}

export interface DashboardOverview {
  greeting: string;
  activeComplaints: number;
  resolvedThisWeek: number;
  averageResponseHours: number;
  attendanceRate: number;
  recentActivity: Activity[];
  priorityBreakdown: BreakdownItem[];
}

export type SearchSuggestionKind = typeof SearchSuggestionKind[keyof typeof SearchSuggestionKind];


export const SearchSuggestionKind = {
  directory: 'directory',
  event: 'event',
  location: 'location',
} as const;

export interface SearchSuggestion {
  id: string;
  label: string;
  kind: SearchSuggestionKind;
  meta: string;
}

export interface TimetableEntry {
  id: string;
  title: string;
  room: string;
  time: string;
  instructor: string;
  color: string;
}

export interface MenuItem {
  id: string;
  meal: string;
  title: string;
  description: string;
  tags: string[];
}

export type ListComplaintsParams = {
status?: ComplaintStatus;
};

export type GetSearchSuggestionsParams = {
/**
 * @minLength 1
 * @maxLength 64
 */
q: string;
};