# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusOS shared API for complaint routing and campus operations
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintCategory } from './complaintCategory';
import type { ComplaintSeverity } from './complaintSeverity';
import type { ComplaintStatus } from './complaintStatus';

export interface Complaint {
  id: string;
  title: string;
  description: string;
  category: ComplaintCategory;
  severity: ComplaintSeverity;
  status: ComplaintStatus;
  location: string;
  createdAt: Date;
  priority: number;
  priorityLabel: string;
}