# lib/api-zod/src/generated/types/createComplaintInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusOS shared API for complaint routing and campus operations
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintCategory } from './complaintCategory';
import type { ComplaintSeverity } from './complaintSeverity';

export interface CreateComplaintInput {
  /**
     * @minLength 5
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  description: string;
  category: ComplaintCategory;
  severity: ComplaintSeverity;
  /**
     * @minLength 2
     * @maxLength 160
     */
  location: string;
}