# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CampusOS shared API for complaint routing and campus operations
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}